extract out interfaces and util
